backend/pyproject.toml
[project]
name = "biostage-backend"
version = "0.1.0"
description = "Telegram post editor bot with admin API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-telegram-bot>=20.8",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "pydantic-settings>=2.2.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
